Fixed up randomness for environment and models.
In general we want to have the largest possible diversity between the
processes, to prevent learning from degenerating.

Randomness in present in the environment through the random seed of the
 atari emulator and the number of no-ops at the beginning of the game.
It is present in the model through the sampling of discrete actions.

This patch makes sure there is a training level random seed, which is saved
to the args.txt file ( even if it has been generated ). This seed is in
turn used to create process level random seeds, which are used for both the
environment and the model. The enviroment random seed is used for the
emulator too.